+++ src/modules/send/send.module.ts
@@ -8,8 +8,11 @@ import { FolderModule } from '../folder/folder.module';
 import { FolderModel } from '../folder/folder.model';
 import { UserModule } from '../user/user.module';
 import { UserModel } from '../user/user.model';
+import { SequelizeSendRepository } from './send-link.repository';
 import { SendController } from './send.controller';
 import { SendUseCases } from './send.usecase';
+import { SendLinkModel } from './send-link.model';
+import { SendLinkItemModel } from './send-link-item.model';
 
 @Module({
   imports: [
